Move changelog-related definitions into their own module
